+++ demo/constructor/src/jsMain/kotlin/constructorJs.kt
@@ -1,0 +1,6 @@
+import space.kscience.visionforge.plotly.PlotlyPlugin
+import space.kscience.visionforge.runVisionClient
+
+public fun main(): Unit = runVisionClient {
+    plugin(PlotlyPlugin)
+}